LibGfx: Move FourCC to its own file
These are used in fonts too, so let's not limit them to ImageLoader.
